Update ADSR to use exponential enveloping
To do this, I model a first order diff-eq to mimic an RC circuit, producing an
exponential envelope that slows down its rate of change. It uses a time constant
of 1/4 the length of the state, in order to hit within 2% of the final goal at
the end of the state.