Updates to v0.5.0.
This version is likely to break projects. It splits out many of the devices
provided in oxcable to a separate crate, as I begin to refocus this crate on
just the signal processing and basic traits. This will allow it to stay leaner,
and only require importing a grab bag of devices if you want to use them.

For the moved devices, see http://github.com/oxcable/basic-devices.

It also has a lot of work on testing behind the scenes. Test coverage is up
significantly, and is more ergonomic. I hope this makes it easier to write
well-tested and bug-free code going forward.

Changelog:
* Moves several devices out into a separate basic-devices crate.
    * ADSR
    * Delay
    * Compressor/Limiter/NoiseGate
    * Reverb
    * Tremolo
* Make wav input/output devices more generic. They operate over anything
  implements Read or Write traits, and can be opened directly by paths, rather
  than already opened files.
* Improves ringbuffer iterators to more closely resemble similar iterators in
  the standard.
* Adds a half-buffered OutputBuffer wrapper for only providing output buffers.
* Greatly improve testing internally with useful tools, including a mocked audio
  device.
* Minor clarity, style and efficiency improvements.
* Documentation tweaks.